Update README.md
Home Maintenance and Interior Design
It refers to the regular work and small repairs that help to ensure that the home is in good working conditions. This involves plumbing, HVAC systems, electrical systems, roofing, among other tasks that may comprise cleaning the gutters, painting, or mowing the lawns. Interior design is the professional service of designing and decorating interiors of spaces to improve functionality and appearance. This comprises choosing colors, accessories, furniture, and illumination in accordance to the desire of the occupants and the functionality of the space, safety concerns among other factors.